Move gameRoutes back into server.js to fix up io.emit issues
